+++ sync-for-payroll/src/codatsyncpayroll/models/shared/metadata.py
@@ -11,7 +11,7 @@ from typing import Optional
 
 @dataclasses.dataclass
 class Metadata:
-    is_deleted: Optional[bool] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('isDeleted'), 'exclude': lambda f: f is None }})
+    is_deleted: Optional[bool] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('isDeleted') }})
     r"""Indicates whether the record has been deleted in the third-party system this record originated from."""